requirements/base.txt
coreapi==2.3.3
django==4.0.8
django-constance==2.9.1
django-cms==3.11.0
djangocms-text-ckeditor==5.1.1
djangocms-file==3.0.0
django-filer==2.2.3
djangorestframework==3.14.0
psycopg2-binary==2.9.5
gunicorn==20.1.0
redis==4.3.4